Add server port (#17)
